+++ src/create-wrapped-apis.ts
@@ -5,6 +5,10 @@ import {
 } from "fake-permissions";
 import { createAPIs } from "./create-apis.js";
 import { createDelegatedGeolocation } from "./delegated-geolocation.js";
+import {
+  createGeolocationObserver,
+  type GeolocationObserver,
+} from "./geolocation-observer.js";
 import { MutableLocationServices } from "./location-services.js";
 import { User } from "./user.js";
 
@@ -28,6 +32,7 @@ export function createWrappedAPIs({
   geolocation: Geolocation;
   isUsingSuppliedAPIs: () => boolean;
   locationServices: MutableLocationServices;
+  observer: GeolocationObserver;
   permissions: Permissions;
   permissionStore: PermissionStore;
   selectAPIs: (useSuppliedAPIs: boolean) => void;
@@ -63,9 +68,12 @@ export function createWrappedAPIs({
       delegates: [fakePermissions, suppliedPermissions],
     });
 
+  const observer = createGeolocationObserver(geolocation, permissions);
+
   return {
     geolocation,
     locationServices,
+    observer,
     permissions,
     permissionStore,
     user,